requirements.txt
SQLAlchemy
pandas
Pillow
pytesseract
streamlit
pytest
numpy
lxml